pyproject.toml
[tool.poetry]
name = "rossum-assignment"
version = "0.0.1"
description = ""
authors= ["Josef Polak"]

[tool.poetry.dependencies]
python = "^3.12"
dict2xml = "^1.7.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"
ruff = "^0.0.263"
 
[tool.poetry.group.test.dependencies]  
pytest-randomly = "^3.15.0"  
pytest = "^8.2.0"  

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100

[tool.ruff]
line-length = 100